backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
supabase==2.3.0

# AI and LLM
openai==1.58.1
langchain==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.28

# External APIs
google-api-python-client==2.115.0
httpx==0.24.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
requests==2.31.0
